Remove service/blit.h, but replace it with something worse
