+++ include/ccf/kv/serialisers/blit_serialiser.h
@@ -3,7 +3,6 @@
 #pragma once
 
 #include "ccf/ds/nonstd.h"
-#include "ccf/entity_id.h"
 #include "ccf/kv/serialisers/serialised_entry.h"
 
 namespace kv::serialisers
@@ -74,11 +73,6 @@ namespace kv::serialisers
       else if constexpr (std::is_same_v<T, std::string>)
       {
         return T(rep.begin(), rep.end());
-      }
-      else if constexpr (nonstd::is_specialization<T, ccf::EntityId>::value)
-      {
-        typename T::Value v(rep.begin(), rep.end());
-        return T(v);
       }
       else
       {
