+++ include/ccf/service/tables/members.h
@@ -5,6 +5,8 @@
 #include "ccf/crypto/pem.h"
 #include "ccf/crypto/sha256_hash.h"
 #include "ccf/entity_id.h"
+#include "ccf/kv/serialisers/blit_serialiser_entity_id.h"
+#include "ccf/kv/serialisers/blit_serialiser_pem.h"
 #include "ccf/service/map.h"
 #include "ccf/service/signed_req.h"
