+++ src/endpoints/authentication/cert_auth.cpp
@@ -6,7 +6,6 @@
 #include "ccf/rpc_context.h"
 #include "ccf/service/tables/members.h"
 #include "ccf/service/tables/users.h"
-#include "service/blit.h"
 #include "service/tables/nodes.h"
 
 namespace ccf
