+++ src/node/node_signature_verify.h
@@ -4,7 +4,6 @@
 
 #include "ccf/crypto/verifier.h"
 #include "ccf/tx.h"
-#include "service/blit.h"
 #include "service/tables/nodes.h"
 
 namespace ccf
