+++ src/node/test/history.cpp
@@ -7,7 +7,6 @@
 #include "kv/store.h"
 #include "kv/test/null_encryptor.h"
 #include "kv/test/stub_consensus.h"
-#include "service/blit.h"
 #include "service/tables/nodes.h"
 #include "service/tables/signatures.h"
